chore: deleting assets files of spotify and souncloud logo. Also renaming components file with maj
